backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
torch==2.2.0
torchvision==0.17.0
pydantic==2.6.3
numpy==1.26.4
pillow==10.2.0
python-multipart==0.0.20